M-Lab: start.sh stop.sh: add check for privileges
Move the check from init scripts to M-Lab scripts: the idea is that the
former are just placeholders for the latter, like I did for MacOSX.

+++ M-Lab/stop.sh
@@ -27,6 +27,12 @@
 #
 
 DEBUG=
+
+if [ `id -u` -ne 0 ]; then
+    echo "$0: FATAL: need root privileges" 1>&2
+    exit 1
+fi
+
 if [ -f /var/run/neubot.pid ]; then
     $DEBUG kill -TERM $($DEBUG cat /var/run/neubot.pid)
 fi
